+++ backend/src/module/seguridad/permiso/entities/permiso.entity.ts
@@ -38,7 +38,7 @@ export class Permiso {
     estado: string;
 
     @Column( {
-        type: 'numeric',
+        type: 'int',
         default: 1,
     } )
     concurrencia: number;
